requirements.txt
gradio>=4.0.0
faiss-cpu
sentence-transformers>=2.2.2
transformers>=4.41.0
accelerate>=0.29.0
torch
torchaudio
